requirements.txt
asgiref==3.3.4
Django==3.2.4
pyaes==1.6.1
Pyrogram==0.18.0
PySocks==1.7.1
python-dotenv==0.17.1
pytz==2021.1
sqlparse==0.4.1
TgCrypto==1.2.2
requests==2.25.1
